Created javadoc

+++ CRS-JEDI-7-Flipkart-POS/src/com/crs/flipkart/dao/GradeCardDaoInterface.java
@@ -11,5 +11,5 @@ import com.crs.flipkart.bean.GradeCard;
  */
 public interface GradeCardDaoInterface {
 	public GradeCard getGradeCard(int studentId);
-	public int getGradeFromCourseId(int studentId, int CourseId);
+	public float getGradeFromCourseId(int studentId, int CourseId);
 }
